// apps/template/src/api/generated/types.gen.ts
export type VoloAbpFeatureManagementFeatureDto = {
    name?: string | null;
    displayName?: string | null;
    value?: string | null;
    provider?: VoloAbpFeatureManagementFeatureProviderDto;
    description?: string | null;
    valueType?: VoloAbpValidationStringValuesIStringValueType;
    depth?: number;
    parentName?: string | null;
};

export type VoloAbpFeatureManagementFeatureGroupDto = {
    name?: string | null;
    displayName?: string | null;
    features?: Array<VoloAbpFeatureManagementFeatureDto> | null;
};

export type VoloAbpFeatureManagementFeatureProviderDto = {
    name?: string | null;
    key?: string | null;
};
// ...
export type VoloAbpValidationStringValuesIStringValueType = {
    readonly name?: string | null;
    readonly properties?: {
        [key: string]: unknown;
    } | null;
    validator?: VoloAbpValidationStringValuesIValueValidator;
};

export type VoloAbpValidationStringValuesIValueValidator = {
    readonly name?: string | null;
    readonly properties?: {
        [key: string]: unknown;
    } | null;
};